rearrange cli commands
