+++ applications/services/brightness_control/cli/cli_command_display.h
