+++ applications/services/brightness_control/cli/cli_command_light_sensor.h
