+++ applications/services/device_info/cli/cli_command_device_info.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <cli/cli_command.h>
+
+void cli_command_device_info(PipeSide* pipe, FuriString* args, void* context);
