+++ applications/services/time/application.fam
@@ -7,4 +7,29 @@ App(
     stack_size=1024 * 2,
     order=150,
     targets=["f20", "f21", "f22"],
+    provides=["cli_date", "cli_timezone"],
+    sources=[
+        "settings/interface_v1.c",
+        "settings/settings.c",
+        "time_update.c",
+        "time.c"
+    ],
+)
+
+App(
+    appid="cli_date",
+    name="date",
+    targets=["f20", "f21", "f22"],
+    apptype=FlipperAppType.CLICMD,
+    entry_point="cli_command_date",
+    sources=["cli/cli_command_date.c"],
+)
+
+App(
+    appid="cli_timezone",
+    name="timezone",
+    targets=["f20", "f21", "f22"],
+    apptype=FlipperAppType.CLICMD,
+    entry_point="cli_command_timezone",
+    sources=["cli/cli_command_timezone.c"],
 )
