+++ applications/services/time/cli/cli_command_date.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <containers/pipe.h>
+
+void cli_command_rtc_date(PipeSide* pipe, FuriString* args, void* context);
