+++ applications/system/updater/cli/cli_command_factory_reset.c
